% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Abalone Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4177 rows and 11 columns.}
\source{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/abalone/abalone.data}
}
\usage{
abalone
}
\description{
Cleaned up abalone dataset.
}
\keyword{datasets}
