% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone_pred}
\alias{abalone_pred}
\title{Abalone Predictions}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 1 columns.}
\usage{
abalone_pred
}
\description{
Predictions on the first 100 rows of the \link{abalone} dataset.
}
\keyword{datasets}
