% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{batch_predict}
\alias{batch_predict}
\title{Batch Predictions from Sagemaker Model}
\usage{
batch_predict(
  object,
  s3_input,
  s3_output,
  instance_count = 1L,
  instance_type = "ml.c4.xlarge",
  wait = TRUE
)
}
\arguments{
\item{object}{The \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}.}

\item{s3_input}{The S3 path to the input object.
The object must be a \code{csv}, with no column names.
The \code{csv} must not contain the outcome column.
Use \link{s3} to construct the path.}

\item{s3_output}{The S3 prefix for the output object or objects.}

\item{instance_count}{The number of instances to run.}

\item{instance_type}{Type of EC2 instance to run. See
\href{https://aws.amazon.com/sagemaker/pricing/instance-types/}{here} for
a list of options and pricing.}

\item{wait}{Boolean that indicates if function should wait to return
until the Sagemaker process is complete.}
}
\value{
S3 output path.

The format is
\code{s3://[s3_output]/[s3_input_object_name].out}
}
\description{
Runs a batch of predictions on Sagemaker model.
The input is a csv file on S3, and predictions are saved
as a csv on S3.
}
