% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.sagemaker}
\alias{predict.sagemaker}
\title{Make Predictions from Sagemaker Model}
\usage{
\method{predict}{sagemaker}(object, new_data)
}
\arguments{
\item{object}{The \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}.}

\item{new_data}{The \code{data.frame} or
\code{\link[tibble:tibble]{tibble::tibble()}} to make new predictions on.
Columns must be in the same order as the training data.
The outcome column must be excluded.}
}
\description{
Returns a vector of predictions from a Sagemaker real-time endpoint.
}
