% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.xgboost.core.Booster}
\alias{predict.xgboost.core.Booster}
\title{Make Predictions Locally}
\usage{
\method{predict}{xgboost.core.Booster}(object, new_data)
}
\arguments{
\item{object}{The \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}.}

\item{new_data}{The \code{data.frame} or
\code{\link[tibble:tibble]{tibble::tibble()}} to make new predictions on.
Columns must be in the same order as the training data.
The outcome column must be excluded.}
}
\description{
This function generics predictions from the
\code{xgboost.core.Booster} object returned
from \code{\link{sagemaker_load_model}}.
}
