% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{read_s3}
\alias{read_s3}
\alias{write_s3}
\title{Read/write \code{csv}s from S3}
\usage{
read_s3(s3_path, delim = ",", col_names = FALSE, ...)

write_s3(x, s3_path, delim = ",", col_names = FALSE, ...)
}
\arguments{
\item{s3_path}{A character vector that forms an S3 path to an object.
Use \link{s3} to construct the path.}

\item{delim}{Single character used to separate fields within a record.}

\item{col_names}{Either \code{TRUE}, \code{FALSE} or a character vector
of column names.

If \code{TRUE}, the first row of the input will be used as the column
names, and will not be included in the data frame. If \code{FALSE}, column
names will be generated automatically: X1, X2, X3 etc.

If \code{col_names} is a character vector, the values will be used as the
names of the columns, and the first row of the input will be read into
the first row of the output data frame.

Missing (\code{NA}) column names will generate a warning, and be filled
in with dummy names \code{X1}, \code{X2} etc. Duplicate column names
will generate a warning and be made unique with a numeric prefix.}

\item{...}{Additional named arguments sent to the underlying API.}

\item{x}{A data frame to write to disk}
}
\description{
Downloads an csv file from S3 and reads it into the R session as a
\code{\link[tibble:tibble]{tibble::tibble()}}.

Writes a \code{tibble} to a S3 object.

Defaults \code{col_names} to \code{FALSE},
because that is \link{batch_predict} and
\link{sagemaker_hyperparameter_tuner} expect.

Interface to \code{\link[readr:read_delim]{readr::read_delim()}}
and \code{\link[readr:format_delim]{readr::format_delim()}}
}
