% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3}
\alias{s3}
\title{Creates S3 Object Paths}
\usage{
s3(...)
}
\arguments{
\item{...}{Characters vectors to combine into an S3 path.}
}
\description{
Returns properly formatted S3 object path,
including the \code{s3://} prefix.
}
\examples{
s3("my_bucket", "prefix", "object_name.csv")
s3(s3_bucket(), "prefix1", "prefix2", "object.csv")

}
