% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_bucket}
\alias{s3_bucket}
\title{Sagemaker Default S3 Bucket}
\usage{
s3_bucket()
}
\description{
Returns the default S3 bucket associated with the Sagemaker session.
Utilizes \code{sagemaker$Session()$default_bucket()}.

Set \code{options(sagemaker.default.bucket = "bucket_name")} to use a
different default bucket.
}
