% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_split}
\alias{s3_split}
\title{Train/Validation Split in S3}
\usage{
s3_split(s3_train, s3_validation)
}
\arguments{
\item{s3_train, s3_validation}{S3 paths to the train/validation datasets.
Construct with \link{s3}.}
}
\description{
Returns the test/validation S3 object paths. The train and validation
data sets should be \code{csv}s with no column names.
}
