% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{sagemaker_attach_tuner}
\alias{sagemaker_attach_tuner}
\title{Attach an Existing Sagemaker Tuning Job}
\usage{
sagemaker_attach_tuner(tuning_job_name)
}
\arguments{
\item{tuning_job_name}{Name of the tuning job, typically something
like \code{"xgboost-191114-2052"}.}
}
\description{
Attaches and loads an existing Sagemaker Tuning Job.
This is useful to analyze a tuning job that was completed in a previous
session.

Returns the same object as \link{sagemaker_hyperparameter_tuner}.
}
