% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{sagemaker_container}
\alias{sagemaker_container}
\alias{sagemaker_xgb_container}
\title{Sagemaker Container}
\usage{
sagemaker_container(container_name, repo_version = "latest", ...)

sagemaker_xgb_container(repo_version = "latest", ...)
}
\arguments{
\item{container_name}{Name of the Sagemaker model container, e.g. "xgboost"}

\item{repo_version}{Version of the model. Defaults to "latest".}

\item{...}{Additional named arguments sent to the underlying API.}
}
\description{
Returns the URI to the Sagemaker model container.
}
\details{
Interface to \code{sagemaker$amazon$amazon_estimator$get_image_uri}.
}
