% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagemaker_operations.R
\name{sagemaker_describe_hyper_parameter_tuning_job}
\alias{sagemaker_describe_hyper_parameter_tuning_job}
\title{Gets a description of a hyperparameter tuning job}
\usage{
sagemaker_describe_hyper_parameter_tuning_job(
  HyperParameterTuningJobName)
}
\arguments{
\item{HyperParameterTuningJobName}{[required] The name of the tuning job to describe.}
}
\description{
Gets a description of a hyperparameter tuning job.
}
\section{Request syntax}{
\preformatted{sagemaker$describe_hyper_parameter_tuning_job(
  HyperParameterTuningJobName = "string"
)
}
}

\keyword{internal}
