% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{sagemaker_estimator}
\alias{sagemaker_estimator}
\alias{sagemaker_xgb_estimator}
\title{Sagemaker Estimator}
\usage{
sagemaker_estimator(
  container,
  train_instance_count = 1L,
  train_instance_type = "ml.m4.xlarge",
  output_path = s3(s3_bucket(), "models/"),
  ...
)

sagemaker_xgb_estimator(
  train_instance_count = 1L,
  train_instance_type = "ml.m4.xlarge",
  output_path = s3(s3_bucket(), "models/"),
  ...
)
}
\arguments{
\item{container}{URI of Sagemaker model container.
See \link{sagemaker_container}.}

\item{train_instance_count}{Number of AWS EC2 instances to train on.}

\item{train_instance_type}{Type of EC2 instance to train on. See
\href{https://aws.amazon.com/sagemaker/pricing/instance-types/}{here} for
a list of options and pricing.}

\item{output_path}{The S3 output path to save the model artifact.}

\item{...}{Additional named arguments sent to the underlying API.}
}
\description{
Sagemaker estimator object. Interface to \code{sagemaker$estimator$Estimator}.
}
