% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{sagemaker_hyperparameter_tuner}
\alias{sagemaker_hyperparameter_tuner}
\title{Start a Sagemaker Hyperparamter Tuning Job}
\usage{
sagemaker_hyperparameter_tuner(
  estimator,
  split,
  hyperparameter_ranges,
  strategy = "Random",
  max_jobs = 10L,
  max_parallel_jobs = 2L,
  ...
)
}
\arguments{
\item{estimator}{Sagemaker estimator from \link{sagemaker_estimator}.}

\item{split}{Train/validation dataset split from \link{s3_split}.}

\item{hyperparameter_ranges}{A named list of model hyperparameters
with \link{sagemaker_ranges} for tuning.}

\item{strategy}{Tuning strategy: \code{"Random"} or \code{"Bayesian"}.}

\item{max_jobs}{Number of unique models to train during tuning.}

\item{max_parallel_jobs}{Number of models to train simultaneously.}

\item{...}{Additional named arguments sent to the underlying API.}
}
\description{
Interface to \code{sagemaker$tuner$HyperparameterTuner}.
}
