% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{sagemaker_integer}
\alias{sagemaker_integer}
\alias{sagemaker_continuous}
\alias{sagemaker_categorical}
\title{Sagemaker Parameter Ranges}
\usage{
sagemaker_integer(min, max, scaling = "Auto")

sagemaker_continuous(min, max, scaling = "Auto")

sagemaker_categorical(values)
}
\arguments{
\item{min}{Minimum value of range.}

\item{max}{Maximum value of range.}

\item{scaling}{\code{"Auto"}, \code{"Linear"}, \code{"Logarithmic"}, or
\code{"ReverseLogarithmic"}.}

\item{values}{List of strings that match categorical parameters
for the Sagemaker estimator.}
}
\description{
Parameter Ranges for Sagemaker hyperparameter tuning.
}
