% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sagemaker_load_model}
\alias{sagemaker_load_model}
\alias{sagemaker_load_model.sagemaker}
\alias{sagemaker_load_model.character}
\title{Loads Model Artifact}
\usage{
sagemaker_load_model(x)

\method{sagemaker_load_model}{sagemaker}(x)

\method{sagemaker_load_model}{character}(x)
}
\arguments{
\item{x}{Either the \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}
or the training job name.}
}
\description{
Loads the model artifact in the current R session.
Currently only supports xgboost models.
}
\details{
xgboost models require the xgboost Python package.
See \link{sagemaker_install_xgboost} to download.

xgboost models return a \code{xgboost.core.Booster}
from the xgboost Python package. See
\code{\link{predict.xgboost.core.Booster}}.
}
