% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{sagemaker_ranges}
\alias{sagemaker_ranges}
\alias{sagemaker_integer}
\alias{sagemaker_continuous}
\alias{sagemaker_categorical}
\title{Sagemaker Parameter Ranges}
\usage{
sagemaker_ranges(
  type = c("integer", "continuous", "categorical"),
  min,
  max,
  scaling = "Auto",
  values = NULL
)

sagemaker_integer(min, max, scaling = "Auto")

sagemaker_continuous(min, max, scaling = "Auto")

sagemaker_categorical(values)
}
\arguments{
\item{type}{Type of range for hyperparameter tuning.}

\item{min}{Minimum value of range.}

\item{max}{Maximum value of range.}

\item{scaling}{\code{"Auto"}, \code{"Linear"}, \code{"Logarithmic"}, or
\code{"ReverseLogarithmic"}.}

\item{values}{List of strings that match categorical parameters
for the \link{sagemaker_estimator}.}
}
\description{
Parameter ranges for Sagemaker hyperparameter tuning.
}
