% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{sagemaker_save_execution_role}
\alias{sagemaker_save_execution_role}
\alias{sagemaker_get_execution_role}
\title{Sagemake Execution Role}
\usage{
sagemaker_save_execution_role(role_arn, profile_name = "sagemaker")

sagemaker_get_execution_role(profile_name = "sagemaker")
}
\arguments{
\item{role_arn}{The ARN of the Sagemaker execution role. See below for
an example.}

\item{profile_name}{Name of profile to save in the \code{config} file.}
}
\description{
Functions that read and write the local Sagemaker execution role.
The role will be saved locally in \code{~/.aws/config}.

Use \code{options(sagemaker.default.role.profile = "profile_name")} to
change the default profile name.
}
\details{
Requires
\href{https://www.viget.com/articles/set-up-aws-cli-and-download-your-s3-files-from-the-command-line/}{awscli}.

The Sagemaker ARN has the following form:

\code{
arn:aws:iam::(account_number):role/service-role/(SageMakerExecutionRole)
}
}
