% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sagemaker_training_job_logs}
\alias{sagemaker_training_job_logs}
\title{Sagemaker Training Job Logs}
\usage{
sagemaker_training_job_logs(job_name)
}
\arguments{
\item{The}{training job name. Typically something like
\code{"xgboost-191114-2052-001-7b33b7a5"}.}
}
\description{
Returns the train/evaluation metrics per round of training for a
training job. Typically associated with \code{nround}s for xgboost
or epochs for neural networks.
}
