% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{sagemaker_deploy_endpoint}
\alias{sagemaker_deploy_endpoint}
\alias{sagemaker_has_endpoint}
\alias{sagemaker_delete_endpoint}
\title{Deploy Sagemaker Endpoint}
\usage{
sagemaker_deploy_endpoint(
  object,
  instance_count = 1L,
  instance_type = "ml.t2.medium",
  wait = TRUE,
  ...
)

sagemaker_has_endpoint(object)

sagemaker_delete_endpoint(object)
}
\description{
Deploy Sagemaker Endpoint
}
