% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sagemaker_download_model}
\alias{sagemaker_download_model}
\title{Downloads Model Artifact}
\usage{
sagemaker_download_model(object, path)
}
\arguments{
\item{object}{The \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}.}

\item{path}{File path to write to.}
}
\description{
Downloads model artifact from S3.
}
