% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sagemaker_download_model}
\alias{sagemaker_download_model}
\alias{sagemaker_download_model.sagemaker}
\alias{sagemaker_download_model.character}
\title{Downloads Model Artifact}
\usage{
sagemaker_download_model(x, path)

\method{sagemaker_download_model}{sagemaker}(x, path)

\method{sagemaker_download_model}{character}(x, path)
}
\arguments{
\item{x}{Either the \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}
or the training job name.}

\item{path}{A file path.}
}
\description{
Downloads model artifact from S3.
}
