% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{sagemaker_estimator}
\alias{sagemaker_estimator}
\alias{sagemaker_xgb_estimator}
\title{Sagemaker Estimator}
\usage{
sagemaker_estimator(
  container,
  instance_count = 1L,
  instance_type = "ml.m4.xlarge",
  s3_output = s3(s3_bucket(), "models/"),
  ...
)

sagemaker_xgb_estimator(
  train_instance_count = 1L,
  train_instance_type = "ml.m4.xlarge",
  s3_output = s3(s3_bucket(), "models/"),
  ...
)
}
\arguments{
\item{container}{URI of Sagemaker model container.
See \link{sagemaker_container}.}

\item{instance_count}{The number of instances to run.}

\item{instance_type}{Type of EC2 instance to run. See
\href{https://aws.amazon.com/sagemaker/pricing/instance-types/}{here} for
a list of options and pricing.}

\item{s3_output}{The S3 output path to save the model artifact.
See \link{s3} to construct the S3 path.}

\item{...}{Additional named arguments sent to the underlying API.}
}
\description{
Sagemaker estimator object. Interface to \code{sagemaker$estimator$Estimator}.
}
