% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{sagemaker_install}
\alias{sagemaker_install}
\title{Installs Sagemaker Dependencies}
\usage{
sagemaker_install(...)
}
\arguments{
\item{...}{Passes named arguments to
\code{\link[reticulate:py_install]{reticulate::py_install()}}.}
}
\description{
Installs the Python package dependencies boto3, sagemaker, and awscli.
}
