% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sagemaker_load_model}
\alias{sagemaker_load_model}
\title{Loads Model Artifact}
\usage{
sagemaker_load_model(object)
}
\arguments{
\item{object}{The \code{sagemaker} model object created by
\link{sagemaker_hyperparameter_tuner} or \link{sagemaker_attach_tuner}.}
}
\description{
Loads the model artifact in the current R session.
Currently only supports xgboost models.
}
\details{
xgboost models require the xgboost Python package.
See \link{sagemaker_install_xgboost} to download.

xgboost models return a \code{xgboost.core.Booster}
from the xgboost Python package. See
\code{\link{predict.xgboost.core.Booster}}.
}
