% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sagemaker_tuning_job_logs}
\alias{sagemaker_tuning_job_logs}
\title{Sagemaker Tuning Job Logs}
\usage{
sagemaker_tuning_job_logs(sagemaker_tuner)
}
\arguments{
\item{sagemaker_tuner}{This is either the \code{sagemaker} object
returned from \link{sagemaker_hyperparameter_tuner} or
\link{sagemaker_attach_tuner} or the name of the tuning job,
typically something like \code{"xgboost-191114-2052"}.}
}
\description{
Returns parameter and validation results from the tuning job.
The result is a data frame has one row per model, along with the trained
validation metric.
}
